backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==6.0.1
PyJWT==2.10.1
google-auth==2.40.3
requests==2.32.5
python-dotenv==1.1.1
mysql-connector-python==9.4.0